requirements.txt
selenium==4.36.0
pytest==8.4.2
Faker==37.12.0
allure-pytest==2.14.1
pytest-xdist==3.6.1
python-dotenv
ruff
dotenv
requests>=2.28.0
lxml>=4.9.0   